move get goal state at sorries to the lean parsing module
